feat: display task progress and estimation
This commit introduces a new feature that displays the progress and estimation of each task.
The progress is calculated as the number of activities completed divided by the total number of activities, and the estimation is calculated as the sum of all activities.
This information is displayed in the task view, and it helps users to track their progress and estimate how much time it will take to complete the task.

Additionally, the commit includes a number of bug fixes and improvements to the user interface.